Split and isolate Keycloak specs
